requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-cors-headers==2.5.2
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
future==0.17.1
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
kombu==4.5.0
oauth2==1.9.0.post1
oauthlib==3.0.1
psycopg2-binary==2.8.1
PyJWT==1.7.1
PySocks==1.6.8
python-twitter==3.5
pytz==2018.9
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tweepy==3.7.0
urllib3==1.24.1
vine==1.3.0
